cpp: bioformats: Move TileCoverage out from tiff namespace
This is also usable in other readers and writers.
